backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
httpx==0.27.0
openai==1.40.0
serpapi==0.1.5
google-search-results==2.4.2
beautifulsoup4==4.12.3
playwright==1.45.0
numpy==1.26.4
redis==5.0.8
python-multipart==0.0.9
Pillow==10.4.0
indic-transliteration==2.3.57

gunicorn==21.2.0
